refactor: Name the shared command module shared (spec 0024)
commands::utils becomes commands::shared, following launchers::shared. The
crate already has a top-level utils module, so the name was taken twice.

Branch: spec-0024
AI-usage: draft (Claude + Opus 5)
